fix typos and light reorganize
