+++ crates/pairwise_nostd_internal/src/lib.rs
@@ -241,13 +241,12 @@ these categories _"Nested"_ and _"Batched"_.
 
 mod apply_points;
 mod bins;
-mod idx_3d_offset;
 mod misc;
 mod parallel;
 mod reduce_utils;
 mod reducer;
-mod spatial;
 mod state;
+mod twopoint;
 
 // I'm not really sure we want to publicly expose reduce_sample, but for now,
 // we expose it to support testing...
@@ -262,5 +261,5 @@ pub use parallel::{
 };
 pub use reduce_utils::reset_full_statepack;
 pub use reducer::{Datum, Histogram, Mean, OutputDescr, Reducer};
-pub use spatial::{CartesianBlock, CellWidth};
 pub use state::{AccumStateView, AccumStateViewMut, StatePackViewMut};
+pub use twopoint::spatial::{CartesianBlock, CellWidth};
