+++ crates/pairwise_nostd_internal/src/twopoint/mod.rs
@@ -1,0 +1,2 @@
+mod idx_3d_offset;
+pub(crate) mod spatial;
