+++ crates/pairwise_nostd_internal/src/twopoint/spatial.rs
